diff group and market creation things
