+++ web/pages/group/loading/[groupId].tsx
@@ -1,8 +1,14 @@
 import { groupPath } from 'common/group'
+import { debounce } from 'lodash'
 import { useRouter } from 'next/router'
-import { useEffect } from 'react'
+import { useEffect, useRef } from 'react'
+import Lottie from 'react-lottie'
+import { Col } from 'web/components/layout/col'
+import { Page } from 'web/components/layout/page'
 import { getGroup } from 'web/lib/supabase/group'
-import { LoadingNewThing } from 'web/pages/loading/[contractId]'
+import loading from '../../../public/lottie/loading-icon.json'
+
+export const LOADING_PING_INTERVAL = 100
 
 export async function getStaticProps(props: { params: { groupId: string } }) {
   const { groupId } = props.params
@@ -20,23 +26,55 @@ export async function getStaticPaths() {
 export default function LoadingNewGroup(props: { groupId: string }) {
   const { groupId } = props
   const router = useRouter()
-  const pingInterval = 200
-  useEffect(() => {
-    const interval = setInterval(waitForGroup, pingInterval) // Adjust the interval as needed
+  const waitForGroup = useRef(
+    debounce(fetchGroup, LOADING_PING_INTERVAL)
+  ).current
 
-    async function waitForGroup() {
-      const newGroup = await getGroup(groupId)
-      if (newGroup) {
-        clearInterval(interval)
-        router.replace(groupPath(newGroup.slug)).catch((e) => {
-          console.log(e)
-        })
-      }
+  async function fetchGroup() {
+    const newGroup = await getGroup(groupId)
+    if (newGroup) {
+      router.replace(groupPath(newGroup.slug)).catch((e) => {
+        console.log(e)
+      })
+    } else {
+      fetchGroup()
     }
-    return () => {
-      clearInterval(interval)
-    }
+  }
+
+  useEffect(() => {
+    waitForGroup()
   }, [])
 
   return <LoadingNewThing thing={'group'} />
+}
+
+export function LoadingNewThing(props: { thing: 'group' | 'contract' }) {
+  const { thing } = props
+  return (
+    <Page>
+      <Col className=" mt-40 w-full items-center justify-center">
+        <Col>
+          <Lottie
+            options={{
+              loop: true,
+              autoplay: true,
+              animationData: loading,
+              rendererSettings: {
+                preserveAspectRatio: 'xMidYMid slice',
+              },
+            }}
+            width={100}
+            isStopped={false}
+            isPaused={false}
+            style={{
+              color: '#6366f1',
+              pointerEvents: 'none',
+              background: 'transparent',
+            }}
+          />
+        </Col>
+        <div className="mt-6">Creating new {thing}...</div>
+      </Col>
+    </Page>
+  )
 }
